Adding automatic PR branch updates.
This will make sure that when a PR is opened, if the base branch is
updated, the PR will be automatically updated as well. This prevents
issues where a merge is attempted and fails because the branch isn't
up-to-date and we have to wait around for it to finish forever.