Adding a bit of wording to the top of the authors page.
Previously, if you clicked on an authors page it listed their bio, then
immediately listed all their articles. This was slightly confusing
(multiple people have brought it up over the last year), so this commit
introduces some simple phrasing that makes it clear that the page you're
on is ONLY a list of the posts this author has written.