move integrator org warnings to after post intros (#1581)
warnings at the start of posts were taking up too much room in post previews in search results.

This moves the warning to above the first section heading in all markdown posts that have the warning.

This moves the warning to above the prerequisites section or the toc in the adoc posts that have those, and just above-the-fold in the one adoc where I saw neither a prerequisites section nor a toc above the fold.